Map Coordinates
Was fiddling with how I could display the map without the 'Fog of War' effect for use as a debug command, and in the process I had the application spit this file out. All it displays are the x and y coordinates of all walkable tiles. Its not formatted worth a rat's ass, but its proof that seeds could be linked with their pathing information and then reconstructed to display identical dungeons for idental seeds, if libtcod doesn't support this feature natively.