Fixed a reverting error >.<
Made a mistake by reverting a previous change. Details of this change are below:
Fixed Forced Dungeon Respawning (no more app restarting to generate a new dungeon)
Added a new debug bind to restart the app ('RCtrl' + 'RAlt' + 'R'). Must be running as Admin with cheats enabled ('RCtrl' + 'RAlt' + '1') and can not be running via a debugger for it to work.

This reverts commit 0140da0055a01f1f4a3be291e8e0e5eb385a5cbe.